refactor: move utils and constants in other directory
